The BP_Test problem
The problem appears to be that for all the screens (BP_Test, BP_Test2 to BP Test7) that they are all of the same type - they all rely on the same Media Player in their BP_Test blueprint (not to be confused with the Level blueprint). Regardless of what changes are done in the level blueprint, they still rely on the same Media Player - therefore all screens play the same thing.