MPT_Base is a problem
MPT_Base is a problem because it takes in one Media Player (Edit -> Details). Previously, this was MP_Base. When you change the Media Player, all screens follow and play that media player as well, because all screens appear to use the same media texture?

I am relatively confident this is THE problem.