Random and different Playback with module operation
Implemented random starting screens with modulo operation such that, with at least 6 media sources in a playlist, that when starting no two media players play the same video.

This was picked after the attempt to read from an array of played media sources (first an array of integer that were appended, then an array of booleans) became a wild goose chase.